% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bic_cat.R
\name{aic_cat}
\alias{aic_cat}
\title{Akaike information criterion for fitted categorical AD models}
\usage{
aic_cat(fit)
}
\arguments{
\item{fit}{A fitted model object of class \code{"cat_fit"} returned by
\code{\link{fit_cat}}.}
}
\value{
A numeric scalar AIC value.
}
\description{
Computes AIC using the fitted log likelihood and a parameter count for
categorical antedependence parameters.
}
\details{
The AIC is computed as:
\deqn{AIC = -2 \times \ell + 2k}
where \eqn{\ell} is the log-likelihood and \eqn{k} is the number of free
parameters.
}
\examples{
set.seed(1)
y <- simulate_cat(40, 5, order = 1, n_categories = 2)
fit <- fit_cat(y, order = 1)
aic_cat(fit)

}
