% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{anova.antedep}
\alias{anova.antedep}
\alias{anova.gau_fit}
\alias{anova.cat_fit}
\alias{anova.inad_fit}
\title{Likelihood-Ratio Comparisons for Nested AD Fits}
\usage{
\method{anova}{gau_fit}(object, ..., test = "Chisq")

\method{anova}{cat_fit}(object, ..., test = "Chisq")

\method{anova}{inad_fit}(object, ..., test = "Chisq")
}
\arguments{
\item{object}{First fitted model object.}

\item{...}{Additional fitted model objects.}

\item{test}{Test label; only \code{"Chisq"} is currently supported.}
}
\value{
An \code{anova} data frame.
}
\description{
Compares one or more nested antedependence model fits in the order supplied.
}
