% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bell.R
\name{Bz}
\alias{Bz}
\title{Bell numbers}
\usage{
Bz(z)
}
\arguments{
\item{z}{nonnegative integer giving the maximum order.}
}
\value{
numeric vector of length \code{z + 1} containing B_0, ..., B_z.
}
\description{
Compute Bell numbers up to order z using a truncated series representation.
}
\details{
This function is for internal use. The results are stored in \code{.Bell_num}
and used by the Bell distribution functions.
}
\keyword{internal}
