% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inad_utils.R
\name{.bell_mean_from_theta}
\alias{.bell_mean_from_theta}
\title{Convert Bell parameter to Bell mean}
\usage{
.bell_mean_from_theta(theta)
}
\description{
Convert Bell parameter to Bell mean
}
\keyword{internal}
