% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inad_utils.R
\name{.bell_theta_from_mean}
\alias{.bell_theta_from_mean}
\title{Convert Bell mean to Bell parameter}
\usage{
.bell_theta_from_mean(mu)
}
\description{
Convert Bell mean to Bell parameter
}
\keyword{internal}
