% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logL_gau_missing.R
\name{.build_gau_covariance}
\alias{.build_gau_covariance}
\title{Build AD covariance matrix from parameters}
\usage{
.build_gau_covariance(order, phi, sigma, n_time)
}
\arguments{
\item{order}{Antedependence order}

\item{phi}{Dependence parameters}

\item{sigma}{Innovation standard deviations}

\item{n_time}{Number of time points}
}
\value{
Covariance matrix (n_time x n_time)
}
\description{
Build AD covariance matrix from parameters
}
\keyword{internal}
