% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_cat.R
\name{.cat_theta_to_prob}
\alias{.cat_theta_to_prob}
\title{Convert unconstrained logits to probability row}
\usage{
.cat_theta_to_prob(theta_row, c)
}
\arguments{
\item{theta_row}{Numeric vector of length c-1}

\item{c}{Number of categories}
}
\value{
Probability vector of length c
}
\description{
Convert unconstrained logits to probability row
}
\keyword{internal}
