% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_inad.R
\name{.ci_alpha_theta_louis_fe}
\alias{.ci_alpha_theta_louis_fe}
\title{Compute alpha and theta CIs using Louis' identity (general version)}
\usage{
.ci_alpha_theta_louis_fe(
  y,
  fit,
  blocks,
  level,
  idx_time,
  ridge,
  thinning,
  innovation
)
}
\description{
Compute alpha and theta CIs using Louis' identity (general version)
}
\keyword{internal}
