% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_inad.R
\name{.ci_wald_i_louis_fe}
\alias{.ci_wald_i_louis_fe}
\title{Wald CI at time i using Louis' identity}
\usage{
.ci_wald_i_louis_fe(
  y,
  i,
  alpha_hat,
  theta_hat,
  tau_hat,
  blocks,
  thinning,
  innovation,
  nb_inno_size = NULL,
  level = 0.95,
  ridge = 0
)
}
\description{
Wald CI at time i using Louis' identity
}
\keyword{internal}
