% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_cat.R
\name{.compute_marginal_ci}
\alias{.compute_marginal_ci}
\title{Compute CIs for marginal parameters}
\usage{
.compute_marginal_ci(marginal, cell_counts, p, c, n_time, n_subjects, z, level)
}
\arguments{
\item{marginal}{Marginal parameter list from fit}

\item{cell_counts}{Cell counts from fit}

\item{p}{Order}

\item{c}{Number of categories}

\item{n_time}{Number of time points}

\item{n_subjects}{Number of subjects}

\item{z}{Z critical value}

\item{level}{Confidence level}
}
\value{
Data frame with CIs
}
\description{
Compute CIs for marginal parameters
}
\keyword{internal}
