% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_cat.R
\name{.compute_transition_ci}
\alias{.compute_transition_ci}
\title{Compute CIs for transition parameters}
\usage{
.compute_transition_ci(transition, cell_counts, p, c, n_time, z, level)
}
\arguments{
\item{transition}{Transition parameter list from fit}

\item{cell_counts}{Cell counts from fit}

\item{p}{Order}

\item{c}{Number of categories}

\item{n_time}{Number of time points}

\item{z}{Z critical value}

\item{level}{Confidence level}
}
\value{
List of data frames with CIs for each time point
}
\description{
Compute CIs for transition parameters
}
\keyword{internal}
