% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrt_mean_ad.R
\name{.construct_ad_covariance}
\alias{.construct_ad_covariance}
\title{Construct AD covariance matrix from fitted model}
\usage{
.construct_ad_covariance(fit, n_time)
}
\arguments{
\item{fit}{Fitted AD model from fit_ad().}

\item{n_time}{Number of time points.}
}
\value{
Covariance matrix.
}
\description{
Construct AD covariance matrix from fitted model
}
\keyword{internal}
