% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrt_mean_gau.R
\name{.construct_gau_covariance}
\alias{.construct_gau_covariance}
\title{Construct AD covariance matrix from fitted model}
\usage{
.construct_gau_covariance(fit, n_time)
}
\arguments{
\item{fit}{Fitted AD model from fit_gau().}

\item{n_time}{Number of time points.}
}
\value{
Covariance matrix.
}
\description{
Construct AD covariance matrix from fitted model
}
\keyword{internal}
