% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ad_utils.R, R/loglik_ad.R
\name{.count_params_ad}
\alias{.count_params_ad}
\title{Count parameters for AD model}
\usage{
.count_params_ad(order, n_time, n_blocks = 1L, estimate_mu = FALSE)
}
\arguments{
\item{order}{Antedependence order}

\item{n_time}{Number of time points}

\item{n_blocks}{Number of blocks (default 1)}

\item{estimate_mu}{Whether mu is estimated.}
}
\value{
Number of free parameters
}
\description{
Count parameters for AD model
}
\keyword{internal}
