% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_utils.R
\name{.count_params_cat}
\alias{.count_params_cat}
\title{Count free parameters for AD(p) categorical model}
\usage{
.count_params_cat(
  order,
  n_categories,
  n_time,
  n_blocks = 1,
  homogeneous = TRUE
)
}
\arguments{
\item{order}{AD order p}

\item{n_categories}{Number of categories c}

\item{n_time}{Number of time points n}

\item{n_blocks}{Number of blocks/groups}

\item{homogeneous}{Whether parameters are shared across blocks}
}
\value{
Number of free parameters
}
\description{
Count free parameters for AD(p) categorical model
}
\keyword{internal}
