% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gau_utils.R
\name{.count_params_gau}
\alias{.count_params_gau}
\title{Count parameters for AD model}
\usage{
.count_params_gau(order, n_time, n_blocks = 1L, estimate_mu = FALSE)
}
\arguments{
\item{order}{Model order.}

\item{n_time}{Number of time points.}

\item{n_blocks}{Number of blocks (1 if no blocks).}

\item{estimate_mu}{Whether mu is estimated.}
}
\value{
Number of free parameters.
}
\description{
Count parameters for AD model
}
\keyword{internal}
