% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrt_homogeneity_inad.R
\name{.count_params_homogeneity}
\alias{.count_params_homogeneity}
\title{Count parameters for homogeneity test}
\usage{
.count_params_homogeneity(
  fit,
  test,
  n_blocks,
  order,
  n_time,
  thinning,
  innovation,
  is_null
)
}
\arguments{
\item{fit}{Fitted model}

\item{test}{Test type}

\item{n_blocks}{Number of groups}

\item{order}{AD order}

\item{n_time}{Number of time points}

\item{thinning}{Thinning type}

\item{innovation}{Innovation type}

\item{is_null}{Whether this is the null model}
}
\value{
Number of free parameters
}
\description{
Count parameters for homogeneity test
}
\keyword{internal}
