% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_cat.R
\name{.default_transition_cat}
\alias{.default_transition_cat}
\title{Create default transition parameters (uniform)}
\usage{
.default_transition_cat(p, c, n_time, n_blocks, homogeneous)
}
\arguments{
\item{p}{Order}

\item{c}{Number of categories}

\item{n_time}{Number of time points}

\item{n_blocks}{Number of groups}

\item{homogeneous}{Whether to share parameters}
}
\value{
List of transition parameters
}
\description{
Create default transition parameters (uniform)
}
\keyword{internal}
