% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_gau_em.R
\name{.em_e_step_gau}
\alias{.em_e_step_gau}
\title{E-step: Compute expected sufficient statistics}
\usage{
.em_e_step_gau(y, order, mu, phi, sigma, blocks, tau)
}
\description{
Computes conditional expectations for missing values and second moments.
}
\keyword{internal}
