% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_gau_em.R
\name{.em_m_step_gau}
\alias{.em_m_step_gau}
\title{M-step: Update parameters from sufficient statistics}
\usage{
.em_m_step_gau(suff_stats, order, blocks, estimate_mu, n_subjects, n_time)
}
\description{
M-step: Update parameters from sufficient statistics
}
\keyword{internal}
