% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_cat.R
\name{.extract_conditional_probs}
\alias{.extract_conditional_probs}
\title{Extract conditional probabilities from transition array}
\usage{
.extract_conditional_probs(trans_array, history)
}
\arguments{
\item{trans_array}{Transition array with last dimension being outcome}

\item{history}{Vector of conditioning values}
}
\value{
Vector of probabilities for current time point
}
\description{
Extract conditional probabilities from transition array
}
\keyword{internal}
