% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_cat.R
\name{.fit_cat_single_pop}
\alias{.fit_cat_single_pop}
\title{Fit CAT model for a single population}
\usage{
.fit_cat_single_pop(y, p, c, n_time, subject_mask = NULL)
}
\arguments{
\item{y}{Data matrix}

\item{p}{Order}

\item{c}{Number of categories}

\item{n_time}{Number of time points}

\item{subject_mask}{Logical vector for subject selection (NULL = all)}
}
\value{
List with marginal, transition, cell_counts, log_l
}
\description{
Fit CAT model for a single population
}
\keyword{internal}
