% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_cat.R
\name{.fit_cat_single_pop_marginalize}
\alias{.fit_cat_single_pop_marginalize}
\title{Fit CAT model with missing data via observed-data likelihood optimization}
\usage{
.fit_cat_single_pop_marginalize(y, p, c, n_time)
}
\arguments{
\item{y}{Data matrix for one population (may contain NA)}

\item{p}{Order}

\item{c}{Number of categories}

\item{n_time}{Number of time points}
}
\value{
List with marginal, transition, cell_counts, log_l, convergence
}
\description{
Fit CAT model with missing data via observed-data likelihood optimization
}
\keyword{internal}
