% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrt_stationarity_cat.R
\name{.fit_cat_stationary_single}
\alias{.fit_cat_stationary_single}
\title{Fit stationary model for single population}
\usage{
.fit_cat_stationary_single(y, p, c, n_time, subject_mask = NULL)
}
\arguments{
\item{y}{Data matrix}

\item{p}{Order}

\item{c}{Number of categories}

\item{n_time}{Number of time points}

\item{subject_mask}{Logical mask for subjects (NULL = all)}
}
\value{
List with marginal, transition, log_l, n_params
}
\description{
Fit stationary model for single population
}
\keyword{internal}
