% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_utils.R
\name{.get_combinations_cat}
\alias{.get_combinations_cat}
\title{Get all category combinations of given length}
\usage{
.get_combinations_cat(n_categories, length)
}
\arguments{
\item{n_categories}{Number of categories c}

\item{length}{Number of positions}
}
\value{
Matrix where each row is a combination (c^length rows, length columns)
}
\description{
Get all category combinations of given length
}
\keyword{internal}
