% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_utils.R
\name{.get_history}
\alias{.get_history}
\title{Extract history for a subject at time k}
\usage{
.get_history(y_row, k, order)
}
\arguments{
\item{y_row}{Single row of data (one subject)}

\item{k}{Current time index}

\item{order}{AD order p}
}
\value{
Integer vector of length min(k-1, order) with category history
}
\description{
Extract history for a subject at time k
}
\keyword{internal}
