% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ad_utils.R
\name{.group_means_ad}
\alias{.group_means_ad}
\title{Compute group-specific means for two-sample case}
\usage{
.group_means_ad(y, blocks)
}
\arguments{
\item{y}{Numeric matrix n_subjects by n_time.}

\item{blocks}{Integer vector indicating group membership.}
}
\value{
List with mu1 and mu2 vectors.
}
\description{
Compute group-specific means for two-sample case
}
\keyword{internal}
