% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik_inad.R
\name{.inad_build_transitions}
\alias{.inad_build_transitions}
\title{Build transition objects for truncated-state missing-data likelihood}
\usage{
.inad_build_transitions(
  order,
  N,
  K,
  alpha1,
  alpha2,
  thinning,
  innovation,
  lambda,
  nb_inno_size
)
}
\description{
Build transition objects for truncated-state missing-data likelihood
}
\keyword{internal}
