% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inad_utils.R
\name{.inad_effective_innovation_mean}
\alias{.inad_effective_innovation_mean}
\title{Effective innovation mean under block effects}
\usage{
.inad_effective_innovation_mean(theta, tau, innovation)
}
\description{
Effective innovation mean under block effects
}
\keyword{internal}
