% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik_inad.R
\name{.inad_make_thin_pmf}
\alias{.inad_make_thin_pmf}
\title{Build truncated thinning pmf for one previous count}
\usage{
.inad_make_thin_pmf(prev, alpha, thinning, K)
}
\description{
Build truncated thinning pmf for one previous count
}
\keyword{internal}
