% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik_inad.R
\name{.inad_order2_dist}
\alias{.inad_order2_dist}
\title{Retrieve/calculate order-2 transition distribution for one previous state pair}
\usage{
.inad_order2_dist(prev2, prev1, info, K)
}
\description{
Retrieve/calculate order-2 transition distribution for one previous state pair
}
\keyword{internal}
