% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik_inad.R
\name{.inad_state_max}
\alias{.inad_state_max}
\title{Heuristic state-space bound for INAD missing-data recursion}
\usage{
.inad_state_max(y, order, alpha1, alpha2, theta, tau)
}
\description{
Heuristic state-space bound for INAD missing-data recursion
}
\keyword{internal}
