% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik_inad.R
\name{.inad_subject_ll_order0}
\alias{.inad_subject_ll_order0}
\title{Subject-level observed-data likelihood for INAD(0)}
\usage{
.inad_subject_ll_order0(y_s, innov_list, K)
}
\description{
Subject-level observed-data likelihood for INAD(0)
}
\keyword{internal}
