% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik_inad.R
\name{.inad_subject_ll_order1}
\alias{.inad_subject_ll_order1}
\title{Subject-level observed-data likelihood for INAD(1)}
\usage{
.inad_subject_ll_order1(y_s, init_pmf, trans1, K)
}
\description{
Subject-level observed-data likelihood for INAD(1)
}
\keyword{internal}
