% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik_inad.R
\name{.inad_subject_ll_order2}
\alias{.inad_subject_ll_order2}
\title{Subject-level observed-data likelihood for INAD(2)}
\usage{
.inad_subject_ll_order2(y_s, init_pmf, trans_t2, order2_info, K)
}
\description{
Subject-level observed-data likelihood for INAD(2)
}
\keyword{internal}
