% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_ad_em.R
\name{.initialize_ad_em}
\alias{.initialize_ad_em}
\title{Initialize parameters for AD EM algorithm}
\usage{
.initialize_ad_em(y, order, blocks, estimate_mu)
}
\description{
Initialize parameters for AD EM algorithm
}
\keyword{internal}
