% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_ad_em.R
\name{.initialize_ad_marginal}
\alias{.initialize_ad_marginal}
\title{Initialize from marginal statistics (ignoring dependence)}
\usage{
.initialize_ad_marginal(y, order, blocks)
}
\description{
Initialize from marginal statistics (ignoring dependence)
}
\keyword{internal}
