% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_gau_em.R
\name{.initialize_gau_em}
\alias{.initialize_gau_em}
\title{Initialize parameters for AD EM algorithm}
\usage{
.initialize_gau_em(y, order, blocks, estimate_mu)
}
\description{
Initialize parameters for AD EM algorithm
}
\keyword{internal}
