% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik_cat.R
\name{.logL_from_counts_cat}
\alias{.logL_from_counts_cat}
\title{Compute log-likelihood from cell counts (faster for large datasets)}
\usage{
.logL_from_counts_cat(y, order, marginal, transition, n_categories)
}
\arguments{
\item{y}{Data matrix}

\item{order}{AD order p}

\item{marginal}{Marginal parameters}

\item{transition}{Transition parameters}

\item{n_categories}{Number of categories}
}
\value{
Scalar log-likelihood
}
\description{
Instead of looping over subjects, this version uses aggregated cell counts.
Equivalent to logL_cat but more efficient.
}
\keyword{internal}
