% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik_inad.R
\name{.logL_inad_missing}
\alias{.logL_inad_missing}
\title{Observed-data INAD likelihood with missing values under MAR}
\usage{
.logL_inad_missing(
  y,
  order,
  thinning,
  innovation,
  alpha,
  theta,
  nb_inno_size = NULL,
  blocks = NULL,
  tau = 0
)
}
\description{
Uses forward recursion over a truncated count state space.
}
\keyword{internal}
