% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik_cat.R
\name{.logL_subject_cat}
\alias{.logL_subject_cat}
\title{Compute log-likelihood contribution from one subject}
\usage{
.logL_subject_cat(y_s, p, c, marginal, transition)
}
\arguments{
\item{y_s}{Integer vector of length n_time (one subject's data)}

\item{p}{Order}

\item{c}{Number of categories}

\item{marginal}{Marginal parameters for this population}

\item{transition}{Transition parameters for this population}
}
\value{
Scalar log-likelihood contribution
}
\description{
Compute log-likelihood contribution from one subject
}
\keyword{internal}
