% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik_cat.R
\name{.logL_subject_cat_marginalize_p2}
\alias{.logL_subject_cat_marginalize_p2}
\title{Observed-data log-likelihood for order-2 CAT model}
\usage{
.logL_subject_cat_marginalize_p2(y_s, c, marginal, transition)
}
\arguments{
\item{y_s}{Subject row (may contain NA)}

\item{c}{Number of categories}

\item{marginal}{Marginal list}

\item{transition}{Transition list}
}
\value{
Scalar log-likelihood contribution
}
\description{
Observed-data log-likelihood for order-2 CAT model
}
\keyword{internal}
