% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik_cat.R
\name{.logsumexp_cat}
\alias{.logsumexp_cat}
\title{Stable log-sum-exp}
\usage{
.logsumexp_cat(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Scalar log(sum(exp(x)))
}
\description{
Stable log-sum-exp
}
\keyword{internal}
