% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_inad.R
\name{.louis_info_i_fe}
\alias{.louis_info_i_fe}
\title{Louis' identity observed information at time i}
\usage{
.louis_info_i_fe(
  y,
  i,
  alpha_hat,
  theta_hat,
  tau_hat,
  blocks,
  thinning,
  innovation,
  nb_inno_size = NULL
)
}
\description{
Louis' identity observed information at time i
}
\keyword{internal}
