% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ad_utils.R
\name{.mle_mean_ad}
\alias{.mle_mean_ad}
\title{Compute MLE of mean vector under AD(p)}
\usage{
.mle_mean_ad(y)
}
\arguments{
\item{y}{Numeric matrix n_subjects by n_time.}
}
\value{
Numeric vector of mean estimates.
}
\description{
Under AD(p), the MLE of mu is the sample mean vector when there are no
covariates.
}
\keyword{internal}
