% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_cat.R
\name{.pack_cat_params}
\alias{.pack_cat_params}
\title{Pack CAT parameters into unconstrained vector}
\usage{
.pack_cat_params(marginal, transition, p, c, n_time)
}
\arguments{
\item{marginal}{Marginal parameter list}

\item{transition}{Transition parameter list}

\item{p}{Order}

\item{c}{Number of categories}

\item{n_time}{Number of time points}
}
\value{
Numeric parameter vector
}
\description{
Pack CAT parameters into unconstrained vector
}
\keyword{internal}
