% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ad_utils.R
\name{.partial_corr_matrix_ad}
\alias{.partial_corr_matrix_ad}
\title{Compute matrix of intervenor-adjusted partial correlations}
\usage{
.partial_corr_matrix_ad(y, mu = NULL)
}
\arguments{
\item{y}{Numeric matrix n_subjects by n_time.}

\item{mu}{Optional mean vector for centering.}
}
\value{
Matrix where entry (i,j) for i > j is r_(i,j|(j+1:i-1)).
}
\description{
Compute matrix of intervenor-adjusted partial correlations
}
\keyword{internal}
