% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ad_utils.R
\name{.psi_kenward}
\alias{.psi_kenward}
\title{Modified psi function for test statistic correction}
\usage{
.psi_kenward(x, y)
}
\arguments{
\item{x}{First argument.}

\item{y}{Second argument.}
}
\value{
Value of psi(x, y).
}
\description{
Computes the psi function used in the modified likelihood ratio test
(Kenward, 1987) for better small-sample approximation.
}
\details{
psi(x, y) is defined as:
\itemize{
\item 0 if x = 0
\item (2y - 1) / (2y(y - 1)) if x = 1
\item 2 * sum from l=1 to x/2 of (y + 2l - 2)^(-1) if x > 0 and even
\item psi(1, y) + psi(x - 1, y + 1) if x > 1 and odd
}
}
\keyword{internal}
