% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_utils.R
\name{.safe_log}
\alias{.safe_log}
\title{Safe log function (0 * log(0) = 0)}
\usage{
.safe_log(x)
}
\arguments{
\item{x}{Numeric vector or array}
}
\value{
Log of x, with 0 for x = 0
}
\description{
Safe log function (0 * log(0) = 0)
}
\keyword{internal}
