% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_cat.R
\name{.simulate_subject_cat}
\alias{.simulate_subject_cat}
\title{Simulate one subject's trajectory}
\usage{
.simulate_subject_cat(n_time, p, c, marginal, transition)
}
\arguments{
\item{n_time}{Number of time points}

\item{p}{Order}

\item{c}{Number of categories}

\item{marginal}{Marginal parameters}

\item{transition}{Transition parameters}
}
\value{
Integer vector of length n_time
}
\description{
Simulate one subject's trajectory
}
\keyword{internal}
