% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_cat.R
\name{.uniform_cat_params}
\alias{.uniform_cat_params}
\title{Build uniform CAT parameter values}
\usage{
.uniform_cat_params(p, c, n_time)
}
\arguments{
\item{p}{Order}

\item{c}{Number of categories}

\item{n_time}{Number of time points}
}
\value{
List with marginal and transition
}
\description{
Build uniform CAT parameter values
}
\keyword{internal}
