% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_cat.R
\name{.unpack_cat_params}
\alias{.unpack_cat_params}
\title{Unpack unconstrained CAT parameter vector}
\usage{
.unpack_cat_params(theta, p, c, n_time)
}
\arguments{
\item{theta}{Numeric parameter vector}

\item{p}{Order}

\item{c}{Number of categories}

\item{n_time}{Number of time points}
}
\value{
List with marginal and transition parameter lists
}
\description{
Unpack unconstrained CAT parameter vector
}
\keyword{internal}
