% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_utils.R
\name{.validate_blocks_cat}
\alias{.validate_blocks_cat}
\title{Validate blocks parameter}
\usage{
.validate_blocks_cat(blocks, n_subjects)
}
\arguments{
\item{blocks}{Block membership vector}

\item{n_subjects}{Number of subjects}
}
\value{
Validated blocks vector (integer, 1-indexed)
}
\description{
Validate blocks parameter
}
\keyword{internal}
