% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_cat.R
\name{.validate_params_cat}
\alias{.validate_params_cat}
\title{Validate transition probability parameters}
\usage{
.validate_params_cat(marginal, transition, p, c, n_time, tol = 1e-08)
}
\arguments{
\item{marginal}{Marginal parameters}

\item{transition}{Transition parameters}

\item{p}{Order}

\item{c}{Number of categories}

\item{n_time}{Number of time points}

\item{tol}{Tolerance for checking sums}
}
\value{
TRUE if valid, otherwise throws error
}
\description{
Checks that transition probabilities are valid (non-negative, sum to 1).
}
\keyword{internal}
