% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_utils.R
\name{.validate_y_cat}
\alias{.validate_y_cat}
\title{Validate categorical data matrix}
\usage{
.validate_y_cat(y, n_categories = NULL)
}
\arguments{
\item{y}{Data matrix}

\item{n_categories}{Number of categories (NULL to infer)}
}
\value{
List with validated y and n_categories
}
\description{
Validate categorical data matrix
}
\keyword{internal}
