% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{fitted.gau_fit}
\alias{fitted.gau_fit}
\alias{residuals.gau_fit}
\title{Fitted Values and Residuals for Gaussian AD Fits}
\usage{
\method{fitted}{gau_fit}(object, y, ...)

\method{residuals}{gau_fit}(object, y, ...)
}
\arguments{
\item{object}{A \code{gau_fit} object.}

\item{y}{Original data matrix.}

\item{...}{Unused.}
}
\value{
A numeric matrix of fitted values or residuals.
}
\description{
Fitted Values and Residuals for Gaussian AD Fits
}
