% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gau_aliases.R
\name{lrt_contrast_gau}
\alias{lrt_contrast_gau}
\title{Linear contrast test for Gaussian AD means (alias of lrt_contrast_ad)}
\usage{
lrt_contrast_gau(...)
}
\arguments{
\item{...}{Passed through to \code{\link{lrt_contrast_ad}}.}
}
\value{
See \code{\link{lrt_contrast_ad}}.
}
\description{
Convenience alias for \code{\link{lrt_contrast_ad}}.
}
\seealso{
\code{\link{lrt_contrast_ad}}
}
