% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gau_aliases.R
\name{lrt_homogeneity_gau}
\alias{lrt_homogeneity_gau}
\title{Covariance homogeneity test for Gaussian AD (alias of lrt_homogeneity_ad)}
\usage{
lrt_homogeneity_gau(...)
}
\arguments{
\item{...}{Passed through to \code{\link{lrt_homogeneity_ad}}.}
}
\value{
See \code{\link{lrt_homogeneity_ad}}.
}
\description{
Convenience alias for \code{\link{lrt_homogeneity_ad}}.
}
\seealso{
\code{\link{lrt_homogeneity_ad}}, \code{\link{lrt_homogeneity_cat}}, \code{\link{lrt_homogeneity_inad}}
}
