% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gau_aliases.R
\name{lrt_one_sample_gau}
\alias{lrt_one_sample_gau}
\title{One-sample mean test for Gaussian AD (alias of lrt_one_sample_ad)}
\usage{
lrt_one_sample_gau(...)
}
\arguments{
\item{...}{Passed through to \code{\link{lrt_one_sample_ad}}.}
}
\value{
See \code{\link{lrt_one_sample_ad}}.
}
\description{
Convenience alias for \code{\link{lrt_one_sample_ad}}.
}
\seealso{
\code{\link{lrt_one_sample_ad}}, \code{\link{lrt_two_sample_ad}}
}
