% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{nobs.antedep}
\alias{nobs.antedep}
\alias{nobs.ad_fit}
\alias{nobs.gau_fit}
\alias{nobs.cat_fit}
\alias{nobs.inad_fit}
\title{Number of Subjects for Antedependence Model Fits}
\usage{
\method{nobs}{ad_fit}(object, ...)

\method{nobs}{gau_fit}(object, ...)

\method{nobs}{cat_fit}(object, ...)

\method{nobs}{inad_fit}(object, ...)
}
\arguments{
\item{object}{A fitted model object.}

\item{...}{Unused.}
}
\value{
Integer scalar: number of subjects.
}
\description{
Number of Subjects for Antedependence Model Fits
}
