% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_methods.R
\name{plot.cat_fit}
\alias{plot.cat_fit}
\title{Plot marginal probabilities of a categorical AD fit}
\usage{
\method{plot}{cat_fit}(x, ...)
}
\arguments{
\item{x}{A \code{cat_fit} object from \code{\link{fit_cat}}.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Displays a stacked bar chart of the estimated marginal category
probabilities at each time point.
}
\examples{
\donttest{
set.seed(1)
y <- simulate_cat(n_subjects = 80, n_time = 5, order = 1, n_categories = 2)
fit <- fit_cat(y, order = 1, n_categories = 2)
plot(fit)
}

}
