% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrt_order_ad.R
\name{print.ad_bic_order}
\alias{print.ad_bic_order}
\title{Print method for BIC order selection}
\usage{
\method{print}{ad_bic_order}(x, ...)
}
\arguments{
\item{x}{Object of class \code{ad_bic_order}.}

\item{...}{Unused.}
}
\description{
Print method for BIC order selection
}
