% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_ad.R
\name{print.ad_ci}
\alias{print.ad_ci}
\title{Print method for AD confidence intervals}
\usage{
\method{print}{ad_ci}(x, ...)
}
\arguments{
\item{x}{An object of class \code{ad_ci}.}

\item{...}{Unused.}
}
\value{
The input object, invisibly.
}
\description{
Print method for AD confidence intervals
}
