% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrt_homogeneity_ad.R
\name{print.ad_homogeneity_test}
\alias{print.ad_homogeneity_test}
\title{Print method for AD homogeneity test}
\usage{
\method{print}{ad_homogeneity_test}(x, ...)
}
\arguments{
\item{x}{Object of class \code{ad_homogeneity_test}.}

\item{...}{Unused.}
}
\description{
Print method for AD homogeneity test
}
