% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrt_order_ad.R
\name{print.ad_order_test}
\alias{print.ad_order_test}
\title{Print method for AD order test}
\usage{
\method{print}{ad_order_test}(x, ...)
}
\arguments{
\item{x}{Object of class \code{ad_order_test}.}

\item{...}{Unused.}
}
\description{
Print method for AD order test
}
