% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_cat.R
\name{print.cat_ci}
\alias{print.cat_ci}
\title{Print method for cat_ci objects}
\usage{
\method{print}{cat_ci}(x, ...)
}
\arguments{
\item{x}{A cat_ci object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for cat_ci objects
}
