% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_utils.R
\name{print.cat_fit}
\alias{print.cat_fit}
\title{Print method for cat_fit objects}
\usage{
\method{print}{cat_fit}(x, ...)
}
\arguments{
\item{x}{A cat_fit object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for cat_fit objects
}
