% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrt_order_cat.R
\name{print.cat_lrt}
\alias{print.cat_lrt}
\title{Print method for cat_lrt objects}
\usage{
\method{print}{cat_lrt}(x, ...)
}
\arguments{
\item{x}{A cat_lrt object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for cat_lrt objects
}
