% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_gau.R
\name{print.gau_ci}
\alias{print.gau_ci}
\title{Print method for AD confidence intervals}
\usage{
\method{print}{gau_ci}(x, ...)
}
\arguments{
\item{x}{An object of class \code{gau_ci}.}

\item{...}{Unused.}
}
\value{
The input object, invisibly.
}
\description{
Print method for AD confidence intervals
}
