% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrt_mean_gau.R
\name{print.gau_contrast_test}
\alias{print.gau_contrast_test}
\title{Print method for AD contrast test}
\usage{
\method{print}{gau_contrast_test}(x, ...)
}
\arguments{
\item{x}{Object of class \code{gau_contrast_test}.}

\item{...}{Unused.}
}
\description{
Print method for AD contrast test
}
