% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_gau.R
\name{print.gau_fit}
\alias{print.gau_fit}
\title{Print Method for gau_fit Objects}
\usage{
\method{print}{gau_fit}(x, ...)
}
\arguments{
\item{x}{A \code{gau_fit} object.}

\item{...}{Additional arguments (ignored).}
}
\value{
The input object, invisibly.
}
\description{
Print Method for gau_fit Objects
}
