% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrt_homogeneity_gau.R
\name{print.gau_homogeneity_test}
\alias{print.gau_homogeneity_test}
\title{Print method for AD homogeneity test}
\usage{
\method{print}{gau_homogeneity_test}(x, ...)
}
\arguments{
\item{x}{Object of class \code{gau_homogeneity_test}.}

\item{...}{Unused.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print method for AD homogeneity test
}
