% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrt_mean_gau.R
\name{print.gau_mean_test}
\alias{print.gau_mean_test}
\title{Print method for AD mean test}
\usage{
\method{print}{gau_mean_test}(x, ...)
}
\arguments{
\item{x}{Object of class \code{gau_mean_test}.}

\item{...}{Unused.}
}
\description{
Print method for AD mean test
}
