% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrt_order_gau.R
\name{print.gau_order_test}
\alias{print.gau_order_test}
\title{Print method for AD order test}
\usage{
\method{print}{gau_order_test}(x, ...)
}
\arguments{
\item{x}{Object of class \code{gau_order_test}.}

\item{...}{Unused.}
}
\description{
Print method for AD order test
}
