% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrt_homogeneity_inad.R
\name{print.homogeneity_tests_inad}
\alias{print.homogeneity_tests_inad}
\title{Print method for homogeneity_tests_inad}
\usage{
\method{print}{homogeneity_tests_inad}(x, digits = 4, ...)
}
\arguments{
\item{x}{Object of class \code{homogeneity_tests_inad}}

\item{digits}{Number of digits for printing}

\item{...}{Unused}
}
\description{
Print method for homogeneity_tests_inad
}
