% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_inad.R
\name{print.inad_ci}
\alias{print.inad_ci}
\title{Print method for INAD confidence intervals}
\usage{
\method{print}{inad_ci}(x, ...)
}
\arguments{
\item{x}{An object of class \code{inad_ci}.}

\item{...}{Unused.}
}
\value{
The input object, invisibly.
}
\description{
Print method for INAD confidence intervals
}
