% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_inad.R
\name{print.inad_fit}
\alias{print.inad_fit}
\title{Print Method for INAD Model Fits}
\usage{
\method{print}{inad_fit}(x, digits = 4, ...)
}
\arguments{
\item{x}{An object of class \code{inad_fit}.}

\item{digits}{Number of digits to print.}

\item{...}{Unused.}
}
\value{
The input object, invisibly.
}
\description{
Print Method for INAD Model Fits
}
