% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrt_homogeneity_inad.R
\name{print.lrt_homogeneity_inad}
\alias{print.lrt_homogeneity_inad}
\title{Print method for lrt_homogeneity_inad}
\usage{
\method{print}{lrt_homogeneity_inad}(x, digits = 4, ...)
}
\arguments{
\item{x}{Object of class \code{lrt_homogeneity_inad}}

\item{digits}{Number of digits for printing}

\item{...}{Unused}
}
\description{
Print method for lrt_homogeneity_inad
}
