% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrt_homogeneity_inad.R
\name{print.test_homogeneity_inad}
\alias{print.test_homogeneity_inad}
\title{Print method for test_homogeneity_inad}
\usage{
\method{print}{test_homogeneity_inad}(x, digits = 4, ...)
}
\arguments{
\item{x}{Object of class \code{test_homogeneity_inad}}

\item{digits}{Number of digits for printing}

\item{...}{Unused}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print method for test_homogeneity_inad
}
