% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_cat.R
\name{summary.cat_ci}
\alias{summary.cat_ci}
\title{Summary method for cat_ci objects}
\usage{
\method{summary}{cat_ci}(object, ...)
}
\arguments{
\item{object}{A cat_ci object}

\item{...}{Additional arguments (ignored)}
}
\value{
A data frame summarizing all CIs
}
\description{
Summary method for cat_ci objects
}
