% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_utils.R
\name{summary.cat_fit}
\alias{summary.cat_fit}
\title{Summary method for cat_fit objects}
\usage{
\method{summary}{cat_fit}(object, ...)
}
\arguments{
\item{object}{A \code{cat_fit} object.}

\item{...}{Unused.}
}
\value{
\code{object}, invisibly.
}
\description{
Summary method for cat_fit objects
}
