% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_gau.R
\name{summary.gau_fit}
\alias{summary.gau_fit}
\title{Summary method for gau_fit objects}
\usage{
\method{summary}{gau_fit}(object, ...)
}
\arguments{
\item{object}{A \code{gau_fit} object.}

\item{...}{Unused.}
}
\value{
\code{object}, invisibly.
}
\description{
Prints a structured table of model-fit statistics and estimated parameters.
}
