% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_inad.R
\name{summary.inad_fit}
\alias{summary.inad_fit}
\title{Summary Method for INAD Model Fits}
\usage{
\method{summary}{inad_fit}(object, ...)
}
\arguments{
\item{object}{An \code{inad_fit} object.}

\item{...}{Unused.}
}
\value{
\code{object}, invisibly.
}
\description{
Summary method for inad_fit objects
}
