% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bic_cat.R
\name{aic_cat}
\alias{aic_cat}
\title{Compute AIC for a categorical antedependence model}
\usage{
aic_cat(fit)
}
\arguments{
\item{fit}{A fitted model object of class \code{"cat_fit"} from \code{fit_cat()}.}
}
\value{
Scalar AIC value.
}
\description{
Calculates the Akaike Information Criterion for a fitted AD(p) model.
}
\details{
The AIC is computed as:
\deqn{AIC = -2 \times \ell + 2k}
where \eqn{\ell} is the log-likelihood and \eqn{k} is the number of free parameters.
}
\examples{
\dontrun{
y <- simulate_cat(100, 5, order = 1, n_categories = 2)
fit <- fit_cat(y, order = 1)
aic_cat(fit)
}

}
