% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gau_aliases.R
\name{aic_gau}
\alias{aic_gau}
\title{Compute AIC for Gaussian AD model (alias of aic_ad)}
\usage{
aic_gau(...)
}
\arguments{
\item{...}{Passed through to \code{\link{aic_ad}}.}
}
\value{
See \code{\link{aic_ad}}.
}
\description{
Convenience alias for \code{\link{aic_ad}}.
}
\seealso{
\code{\link{aic_ad}}, \code{\link{aic_cat}}, \code{\link{aic_inad}}
}
