% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{as.ts.antedep_sim}
\alias{as.ts.antedep_sim}
\alias{as.ts.gau_sim}
\alias{as.ts.cat_sim}
\alias{as.ts.inad_sim}
\title{Convert Antedependence Simulation Output to a Time-Series Object}
\usage{
\method{as.ts}{gau_sim}(x, ...)

\method{as.ts}{cat_sim}(x, ...)

\method{as.ts}{inad_sim}(x, ...)
}
\arguments{
\item{x}{A matrix returned by \code{\link{simulate_gau}},
\code{\link{simulate_cat}}, or \code{\link{simulate_inad}}.}

\item{...}{Additional arguments passed to \code{\link[stats]{ts}}.}
}
\value{
A \code{ts} or \code{mts} object.
}
\description{
Convert Antedependence Simulation Output to a Time-Series Object
}
