% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gau_aliases.R
\name{bic_gau}
\alias{bic_gau}
\title{Compute BIC for Gaussian AD model (alias of bic_ad)}
\usage{
bic_gau(...)
}
\arguments{
\item{...}{Passed through to \code{\link{bic_ad}}.}
}
\value{
See \code{\link{bic_ad}}.
}
\description{
Convenience alias for \code{\link{bic_ad}}.
}
\seealso{
\code{\link{bic_ad}}, \code{\link{bic_cat}}, \code{\link{bic_inad}}
}
