% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gau_aliases.R
\name{bic_order_gau}
\alias{bic_order_gau}
\title{Compare Gaussian AD orders by BIC (alias of bic_order_ad)}
\usage{
bic_order_gau(...)
}
\arguments{
\item{...}{Passed through to \code{\link{bic_order_ad}}.}
}
\value{
See \code{\link{bic_order_ad}}.
}
\description{
Convenience alias for \code{\link{bic_order_ad}}.
}
\seealso{
\code{\link{bic_order_ad}}, \code{\link{bic_order_cat}}, \code{\link{bic_order_inad}}
}
