% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrt_order_inad.R
\name{bic_order_inad}
\alias{bic_order_inad}
\title{BIC Model Order Comparison}
\usage{
bic_order_inad(
  y,
  max_order = 2,
  thinning = "binom",
  innovation = "pois",
  blocks = NULL,
  ...
)
}
\arguments{
\item{y}{Integer matrix.}

\item{max_order}{Maximum order (1 or 2).}

\item{thinning}{Thinning operator.}

\item{innovation}{Innovation distribution.}

\item{blocks}{Optional block assignments.}

\item{...}{Additional arguments.}
}
\value{
A list with class "bic_order_inad".
}
\description{
BIC Model Order Comparison
}
