% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gau_aliases.R
\name{ci_gau}
\alias{ci_gau}
\title{Confidence intervals for Gaussian AD model (alias of ci_ad)}
\usage{
ci_gau(...)
}
\arguments{
\item{...}{Passed through to \code{\link{ci_ad}}.}
}
\value{
See \code{\link{ci_ad}}.
}
\description{
Convenience alias for \code{\link{ci_ad}}.
}
\seealso{
\code{\link{ci_ad}}, \code{\link{ci_cat}}, \code{\link{ci_inad}}
}
