% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{deviance.antedep}
\alias{deviance.antedep}
\alias{deviance.ad_fit}
\alias{deviance.gau_fit}
\alias{deviance.cat_fit}
\alias{deviance.inad_fit}
\title{Deviance for Antedependence Model Fits}
\usage{
\method{deviance}{ad_fit}(object, ...)

\method{deviance}{gau_fit}(object, ...)

\method{deviance}{cat_fit}(object, ...)

\method{deviance}{inad_fit}(object, ...)
}
\arguments{
\item{object}{A fitted model object.}

\item{...}{Unused.}
}
\value{
A scalar deviance value.
}
\description{
Deviance for Antedependence Model Fits
}
