% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gau_aliases.R
\name{em_gau}
\alias{em_gau}
\title{EM algorithm for Gaussian AD model (alias of em_ad)}
\usage{
em_gau(...)
}
\arguments{
\item{...}{Passed through to \code{\link{em_ad}}.}
}
\value{
See \code{\link{em_ad}}.
}
\description{
Convenience alias for \code{\link{em_ad}}.
}
\seealso{
\code{\link{em_ad}}, \code{\link{em_inad}}
}
