% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gau_aliases.R
\name{fit_gau}
\alias{fit_gau}
\title{Fit Gaussian AD model (alias of fit_ad)}
\usage{
fit_gau(...)
}
\arguments{
\item{...}{Passed through to \code{\link{fit_ad}}.}
}
\value{
See \code{\link{fit_ad}}.
}
\description{
Convenience alias for \code{\link{fit_ad}}.
}
\seealso{
\code{\link{fit_ad}}, \code{\link{fit_cat}}, \code{\link{fit_inad}}
}
