% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gau_aliases.R
\name{logL_gau}
\alias{logL_gau}
\title{Log-likelihood for Gaussian AD model (alias of logL_ad)}
\usage{
logL_gau(...)
}
\arguments{
\item{...}{Passed through to \code{\link{logL_ad}}.}
}
\value{
See \code{\link{logL_ad}}.
}
\description{
Convenience alias for \code{\link{logL_ad}}.
}
\seealso{
\code{\link{logL_ad}}, \code{\link{logL_cat}}, \code{\link{logL_inad}}
}
