% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gau_aliases.R
\name{lrt_order_gau}
\alias{lrt_order_gau}
\title{LRT for Gaussian AD order (alias of lrt_order_ad)}
\usage{
lrt_order_gau(...)
}
\arguments{
\item{...}{Passed through to \code{\link{lrt_order_ad}}.}
}
\value{
See \code{\link{lrt_order_ad}}.
}
\description{
Convenience alias for \code{\link{lrt_order_ad}}.
}
\seealso{
\code{\link{lrt_order_ad}}, \code{\link{lrt_order_cat}}, \code{\link{lrt_order_inad}}
}
