% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{nobs.antedep}
\alias{nobs.antedep}
\alias{nobs.gau_fit}
\alias{nobs.cat_fit}
\alias{nobs.inad_fit}
\title{Number of observations for antedependence model fits}
\usage{
\method{nobs}{gau_fit}(object, ...)

\method{nobs}{cat_fit}(object, ...)

\method{nobs}{inad_fit}(object, ...)
}
\arguments{
\item{object}{A fitted model object of class \code{gau_fit},
\code{cat_fit}, or \code{inad_fit}.}

\item{...}{Unused.}
}
\value{
Integer scalar: number of subjects.
}
\description{
Number of observations for antedependence model fits
}
