% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gau_aliases.R
\name{simulate_gau}
\alias{simulate_gau}
\title{Simulate Gaussian AD data (alias of simulate_ad)}
\usage{
simulate_gau(...)
}
\arguments{
\item{...}{Passed through to \code{\link{simulate_ad}}.}
}
\value{
See \code{\link{simulate_ad}}.
}
\description{
Convenience alias for \code{\link{simulate_ad}}.
}
\seealso{
\code{\link{simulate_ad}}, \code{\link{simulate_cat}}, \code{\link{simulate_inad}}
}
